requirements.txt
anthropic == 0.75.0

# Google AI APIs
google-genai == 1.57.0

# Async support
aiofiles == 25.1.0

# Video processing
ffmpeg-python == 0.2.0

# Data validation
pydantic == 2.12.5

# File handling
Pillow == 12.1.0

# Utilities
python-dotenv == 1.2.1
typer == 0.21.1
rich == 14.2.0

# Logging
structlog == 25.5.0
